fix(lang/py): componentize-py update
This commit fixes the discrepancy due to changes in how
`componentize-py` works that have been merged recently.

Both bindings generation and the generated code have changed such that
the guide is no longer correct as to what to import and implement --
the new docs prefer a `wit_world` import which is generated.