fix(lang/py): componentize py usage (#296)
* fix(lang/py): componentize-py update

This commit fixes the discrepancy due to changes in how
`componentize-py` works that have been merged recently.

Both bindings generation and the generated code have changed such that
the guide is no longer correct as to what to import and implement --
the new docs prefer a `wit_world` import which is generated.

* fix(lang/py): remove section regarding exporting interface

This commit removes the section regarding exporting
interfaces (instead of functions) as the code was long ago changed to
use interfaces by default.

* refactor(lang/py): use markdown include for WIT file

* fix(lang/py): exports interface in python docs

Co-authored-by: Wolfgang Meier <womeier@posteo.de>

---------

Co-authored-by: Wolfgang Meier <womeier@posteo.de>