fix: Replace instant with datetime in wit-example.md (#302)
I would imagine what happened here is that what is currently defined as datetime used to be called instant. It looks like it got renamed in some places, but not everywhere.

This patch tries to take care of that.

Apologies if I'm messing something up, this is my first pull request ever in this project.